test: :white_check_mark: перенести round-trip форм в интеграционные
Массовая проверка XML-фикстур соединяет разбор, правила, индекс конфигурации и сериализацию, поэтому не должна учитываться в лимите модульных тестов.
